console,oauth: Move common store logic to global store lib
