+++ pkg/webui/console/store/actions/organizations.js
@@ -13,13 +13,13 @@
 // limitations under the License.
 
 import createRequestActions from '@ttn-lw/lib/store/actions/create-request-actions'
-
 import {
   createPaginationRequestActions,
   createPaginationBaseActionType,
   createPaginationDeleteBaseActionType,
   createPaginationDeleteActions,
-} from './pagination'
+} from '@ttn-lw/lib/store/actions/pagination'
+
 import {
   startEventsStream,
   createStartEventsStreamActionType,
