+++ pkg/webui/console/store/selectors/identity-server.js
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { GET_IS_CONFIGURATION_BASE } from '@console/store/actions/identity-server'
+import { createFetchingSelector } from '@ttn-lw/lib/store/selectors/fetching'
+import { createErrorSelector } from '@ttn-lw/lib/store/selectors/error'
 
-import { createFetchingSelector } from './fetching'
-import { createErrorSelector } from './error'
+import { GET_IS_CONFIGURATION_BASE } from '@console/store/actions/identity-server'
 
 const selectIsStore = state => state.is
