+++ pkg/webui/console/store/selectors/users.js
@@ -12,18 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import { createFetchingSelector } from '@ttn-lw/lib/store/selectors/fetching'
+import { createErrorSelector } from '@ttn-lw/lib/store/selectors/error'
+import {
+  createPaginationIdsSelectorByEntity,
+  createPaginationTotalCountSelectorByEntity,
+} from '@ttn-lw/lib/store/selectors/pagination'
+
 import {
   GET_USERS_LIST_BASE,
   GET_USER_BASE,
   GET_USER_RIGHTS_LIST_BASE,
 } from '@console/store/actions/users'
 
-import { createFetchingSelector } from './fetching'
-import { createErrorSelector } from './error'
-import {
-  createPaginationIdsSelectorByEntity,
-  createPaginationTotalCountSelectorByEntity,
-} from './pagination'
 import { createRightsSelector, createPseudoRightsSelector } from './rights'
 
 const ENTITY = 'users'
