+++ pkg/webui/console/store/selectors/webhook-templates.js
@@ -12,13 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import { createFetchingSelector } from '@ttn-lw/lib/store/selectors/fetching'
+import { createErrorSelector } from '@ttn-lw/lib/store/selectors/error'
+
 import {
   LIST_WEBHOOK_TEMPLATES_BASE,
   GET_WEBHOOK_TEMPLATE_BASE,
 } from '@console/store/actions/webhook-templates'
-
-import { createFetchingSelector } from './fetching'
-import { createErrorSelector } from './error'
 
 const selectWebhookTemplatesStore = state => state.webhookTemplates
 const selectWebhookTemplatesEntitiesStore = state => selectWebhookTemplatesStore(state).entities
