+++ pkg/webui/lib/store/reducers/pagination.js
@@ -1,4 +1,4 @@
-// Copyright © 2019 The Things Network Foundation, The Things Industries B.V.
+// Copyright © 2020 The Things Network Foundation, The Things Industries B.V.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,7 +17,7 @@ import {
   createPaginationRequestActions,
   createPaginationDeleteActions,
   createPaginationByIdDeleteActions,
-} from '@console/store/actions/pagination'
+} from '@ttn-lw/lib/store/actions/pagination'
 
 const defaultState = {
   ids: [],
