+++ pkg/webui/lib/store/selectors/error.js
